+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Change
 
-- Added enableExpand option to Accordion
+- Add isCollapsible option to Accordion
 
 # [0.20.1] - 12-07-2019